requirements.txt
# Core dependencies - проверенные и стабильные версии
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0

# WebSocket и real-time коммуникация
python-socketio==5.10.0
websockets==12.0

# Redis для межсервисной коммуникации
redis[hiredis]==5.0.1
aioredis==2.0.1

# AI и NLP
google-generativeai==0.3.2

# Voice2Voice PRODUCTION-READY компоненты
openai-whisper==20231117
elevenlabs==0.2.27

# Аудио обработка
numpy>=1.26.0
soundfile==0.12.1
pydub==0.25.1
torch>=2.0.0
webrtcvad==2.0.10  # Voice Activity Detection для оптимизации

# HTTP клиенты для API интеграции
httpx[http2]==0.25.2
aiohttp==3.9.1

# Логирование и мониторинг
structlog==23.2.0
prometheus-client==0.19.0

# Утилиты
python-dotenv==1.0.0
pyyaml==6.0.1
uuid7==0.1.0

# Тестирование
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0

# Development tools
black==23.12.0
isort==5.13.2
mypy==1.7.1